Put ShiftedArrowKeyHandler to ShiftedSpecialKeyHandler
