+++ src/com/maddyhome/idea/vim/handler/specialkeys/ShiftedSpecialKeyHandler.kt
@@ -38,8 +38,8 @@ import com.maddyhome.idea.vim.option.Options
  */
 abstract class ShiftedSpecialKeyHandler : EditorActionHandlerBase() {
     final override fun execute(editor: Editor, context: DataContext, cmd: Command): Boolean {
-        val keymodelOption = Options.getInstance().getListOption(Options.KEYMODEL)
-        val startSel = keymodelOption?.contains("startsel") == true
+        val keymodelOption = Options.getInstance().getListOption(Options.KEYMODEL)!!
+        val startSel = "startsel" in keymodelOption
         if (startSel && !CommandState.inVisualMode(editor) && !CommandState.inSelectMode(editor)) {
             if (Options.getInstance().getListOption(Options.SELECTMODE)?.contains("key") == true) {
                 VimPlugin.getVisualMotion().enterSelectMode(editor, CommandState.SubMode.VISUAL_CHARACTER)
@@ -57,4 +57,48 @@ abstract class ShiftedSpecialKeyHandler : EditorActionHandlerBase() {
      * or contains one of `continue*` values but in different mode.
      */
     abstract fun motion(editor: Editor, context: DataContext, cmd: Command)
+}
+
+/**
+ * @author Alex Plate
+ *
+ * Handler for SHIFTED arrow keys
+ *
+ * This handler is used to properly handle there keys according to current `keymodel` and `selectmode` options
+ *
+ * Handler is called once for all carets
+ */
+abstract class ShiftedArrowKeyHandler : EditorActionHandlerBase() {
+    final override fun execute(editor: Editor, context: DataContext, cmd: Command): Boolean {
+        val keymodelOption = Options.getInstance().getListOption(Options.KEYMODEL)!!
+        val startSel = "startsel" in keymodelOption
+        val aContinueSelect = "acontinueselect" in keymodelOption
+        val aContinueVisual = "acontinuevisual" in keymodelOption
+        val inVisualMode = CommandState.inVisualMode(editor)
+        val inSelectMode = CommandState.inSelectMode(editor)
+        if (startSel || aContinueSelect && inSelectMode || aContinueVisual && inVisualMode) {
+            if (!inVisualMode && !inSelectMode) {
+                if ("key" in Options.getInstance().getListOption(Options.SELECTMODE)!!) {
+                    VimPlugin.getVisualMotion().enterSelectMode(editor, CommandState.SubMode.VISUAL_CHARACTER)
+                } else {
+                    VimPlugin.getVisualMotion()
+                            .toggleVisual(editor, 1, 0, CommandState.SubMode.VISUAL_CHARACTER)
+                }
+            }
+            motionWithKeyModel(editor, context, cmd)
+        } else {
+            motionWithoutKeyModel(editor, context, cmd)
+        }
+        return true
+    }
+
+    /**
+     * This method is called when `keymodel` contains `startsel`, or one of `continue*` values in corresponding mode
+     */
+    abstract fun motionWithKeyModel(editor: Editor, context: DataContext, cmd: Command)
+    /**
+     * This method is called when `keymodel` doesn't contain `startsel`,
+     * or contains one of `continue*` values but in different mode.
+     */
+    abstract fun motionWithoutKeyModel(editor: Editor, context: DataContext, cmd: Command)
 }
